feat(model): implement server ping as timeout
By pinging the target server, we are able to workaround a problem with
Chrome for Android taking ~1 min to timeout when the remote server
is offline.
So, if the ping (which is a HEAD request) takes more than the configured
timeout, all requests will be aborted assuming that the target server is
unreachable. Therefore, the cache is hit faster.