chore(): modularize all the things!
This commit introduces several changes to the model service.
Mainly, lots of stuff have been removed from inside of it to minor
pieces where they can be better tested, for example.

With this increase of modularization, the future modelSync service
will be able to make use of the same low level parts of the model
service, namely the requests and standardized PouchDB instances.