feat(model, $modelRequest): move ID field responsability
Problem described in #25: The RESTful API was responsible for
returning in a header what the ID field was in the response body.
Now, we're alleviating this problem by making the client-side the only
responsible for knowing the ID fields of each resource.

BREAKING CHANGE: The previously existing field
$modelRequestProvider.idFieldHeader has been removed so that the
option is read only from the Model constructor.