fix(model): re-read HTTP method in .update() and .patch()
Those methods use a factory function to be built. Until now, the HTTP
method used was passed only while building them.
This implied that the HTTP method would never be changed -- which
is generally the case in production environments. However, it's a very
trivial change to enhance our tests, which won't hurt anyone!